Update - ParseInt to parseFloat for Monthly Invoice Number

+++ mp_cl_create_service_change.js
@@ -705,28 +705,28 @@ function saveRecord() {
 
             if (monday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 1;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
 
             if (tuesday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 2;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
             if (wednesday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 3;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
             if (thursday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 4;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
             if (friday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 5;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
             if (adhoc_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 6;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
 
             var row_service_change_id = edit_class_elem[i].getAttribute('data-servicechangeid');
@@ -990,27 +990,27 @@ function saveRecord() {
 
             if (monday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 1;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
             if (tuesday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 2;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
             if (wednesday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 3;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
             if (thursday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 4;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
             if (friday_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 5;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
             if (adhoc_class_elem[i].checked == true) {
                 freqArray[freqArray.length] = 6;
-                monthly_service_rate = monthly_service_rate + parseInt(new_service_price_class_elem[i].value);
+                monthly_service_rate = monthly_service_rate + parseFloat(new_service_price_class_elem[i].value);
             }
 
             var row_service_change_id = edit_class_elem[i].getAttribute('data-servicechangeid');
